23.8.17-A1/23.8.17-A1/bot/requirements.txt
Pyrogram
TgCrypto
python-dotenv
speedtest-cli
telegraph[aio]
pyrate_limiter 
cachetools
aiofiles
psutil
httpx
motor
bs4
uvloop
pillow

beautifulsoup4
aiomysql
numexpr
aiohttp
fastapi
APScheduler